Making Websites With October CMS - Part 47 - Better Links and URLs

+++ plugins/rainlab/user/updates/users_add_deleted_at.php
@@ -15,9 +15,11 @@ class UsersAddDeletedAt extends Migration
 
     public function down()
     {
-        Schema::table('users', function($table)
-        {
-            $table->dropColumn('deleted_at');
-        });
+        if (Schema::hasColumn('users', 'deleted_at')) {
+            Schema::table('users', function($table)
+            {
+                $table->dropColumn('deleted_at');
+            });
+        }
     }
 }
