+++ plugins/rainlab/user/updates/users_add_login_column.php
@@ -30,10 +30,11 @@ class UsersAddLoginColumn extends Migration
 
     public function down()
     {
-        Schema::table('users', function($table)
-        {
-            $table->dropColumn('login');
-        });
+        if (Schema::hasColumn('users', 'login')) {
+            Schema::table('users', function($table)
+            {
+                $table->dropColumn('login');
+            });
+        }
     }
-
 }
