+++ plugins/rainlab/user/updates/users_add_surname.php
@@ -15,9 +15,11 @@ class UsersAddSurname extends Migration
 
     public function down()
     {
-        Schema::table('users', function($table)
-        {
-            $table->dropColumn('surname');
-        });
+        if (Schema::hasColumn('users', 'surname')) {
+            Schema::table('users', function($table)
+            {
+                $table->dropColumn('surname');
+            });
+        }
     }
 }
